refactor config parser
